+++ src/components/Checkbox/Checkbox.stories.tsx
@@ -1,6 +1,9 @@
-import type { StoryObj } from "@storybook/react";
+import type { StoryObj, Meta } from "@storybook/react";
 import React from "react";
-import Checkbox, { CheckboxInput, Label } from "./Checkbox";
+import CheckboxInput from "../CheckboxInput";
+import CheckboxLabel from "../CheckboxLabel";
+import { Checkbox } from "./Checkbox";
+import styles from "./Checkbox.stories.module.css";
 
 const defaultArgs = {
   disabled: false,
@@ -18,14 +21,8 @@ export default {
       control: "radio",
       options: [true, false, "indeterminate"],
     },
-    size: {
-      description: "Size of the checkbox label.",
-      control: "radio",
-      options: ["small", "medium"],
-      table: { defaultValue: { summary: "medium" } },
-    },
   },
-};
+} as Meta<Args>;
 
 type Args = React.ComponentProps<typeof Checkbox>;
 
@@ -50,12 +47,12 @@ export const Checked: StoryObj<Args> = {
 export const Small: StoryObj<Args> = {
   ...Default,
   args: {
-    size: "small",
+    size: "sm",
   },
 };
 
 export const SmallChecked: StoryObj<Args> = {
-  render: () => <CheckedExample size="small" {...defaultArgs} />,
+  render: () => <CheckedExample size="sm" {...defaultArgs} />,
 };
 
 export const Indeterminate: StoryObj<Args> = {
@@ -68,7 +65,7 @@ export const Indeterminate: StoryObj<Args> = {
 
 export const Disabled = {
   render: () => (
-    <table className="border-separate" style={{ borderSpacing: "2rem" }}>
+    <table className={styles["disabled--table"]}>
       <tbody>
         {[false, true, "indeterminate" as const].map((checked, i) => (
           <tr key={i}>
@@ -84,8 +81,9 @@ export const Disabled = {
     </table>
   ),
   parameters: {
-    // mainly for demonstration purposes
-    snapshot: { skip: true },
+    axe: {
+      disabledRules: ["color-contrast"],
+    },
   },
 };
 
@@ -95,6 +93,16 @@ export const WithoutVisibleLabel: StoryObj<Args> = {
     "aria-label": "a checkbox has no name",
     label: undefined,
   },
+  render: (args) => (
+    <>
+      <Checkbox {...args} />
+      <Checkbox {...args} checked readOnly />
+      <Checkbox {...args} checked="indeterminate" readOnly />
+      <Checkbox {...args} disabled />
+      <Checkbox {...args} checked disabled />
+      <Checkbox {...args} checked="indeterminate" disabled />
+    </>
+  ),
 };
 
 export const LongLabels = {
@@ -102,18 +110,89 @@ export const LongLabels = {
     const label = "Lorem ipsum dolor sit amet, consectetur adipiscing elit";
 
     return (
-      <div className="w-80 grid grid-cols-2 gap-4">
+      <div className={styles["longlabels--grid"]}>
         <Checkbox label={label} />
-        <Checkbox label={label} size="small" />
+        <Checkbox label={label} size="sm" />
+        <Checkbox disabled label={label} />
+        <Checkbox disabled label={label} size="sm" />
       </div>
     );
+  },
+  parameters: {
+    axe: {
+      disabledRules: ["color-contrast"],
+    },
+  },
+};
+
+// For visual regression testing
+export const LabelsOnly: StoryObj = {
+  render: (args) => (
+    <>
+      <CheckboxLabel {...args} htmlFor="md-label" text="Medium label" />
+      <br />
+      <CheckboxLabel
+        {...args}
+        htmlFor="sm-label"
+        size="sm"
+        text="Small label"
+      />
+      <br />
+      <CheckboxLabel
+        {...args}
+        htmlFor="long-md-label"
+        text="Long md label lorem ipsum dolor sit amet, consectetur adipiscing elit"
+      />
+      <br />
+      <CheckboxLabel
+        {...args}
+        htmlFor="long-sm-label"
+        size="sm"
+        text="Long sm label lorem ipsum dolor sit amet, consectetur adipiscing elit"
+      />
+      <br />
+      <CheckboxLabel
+        {...args}
+        disabled
+        htmlFor="md-disabled-label"
+        text="Medium disabled"
+      />
+      <br />
+      <CheckboxLabel
+        {...args}
+        disabled
+        htmlFor="sm-disabled-label"
+        size="sm"
+        text="Small disabled"
+      />
+      <br />
+      <CheckboxLabel
+        {...args}
+        disabled
+        htmlFor="long-md-disabled-label"
+        text="Long md disabled label lorem ipsum dolor sit amet, consectetur adipiscing elit"
+      />
+      <br />
+      <CheckboxLabel
+        {...args}
+        disabled
+        htmlFor="long-sm-disabled-label"
+        size="sm"
+        text="Long sm disabled label lorem ipsum dolor sit amet, consectetur adipiscing elit"
+      />
+    </>
+  ),
+  parameters: {
+    axe: {
+      disabledRules: ["color-contrast"],
+    },
   },
 };
 
 export const WithCustomPositioning = {
   render: () => (
     <span>
-      <Label htmlFor="test">label on left</Label>
+      <CheckboxLabel htmlFor="test" text="label on left" />
       <CheckboxInput id="test" />
     </span>
   ),