+++ src/components/Checkbox/index.ts
@@ -1,2 +1,2 @@
-export { default, CheckboxInput, Label } from "./Checkbox";
+export { Checkbox as default } from "./Checkbox";
 export type { CheckboxProps } from "./Checkbox";