+++ src/components/CheckboxLabel/index.ts
@@ -1,0 +1,2 @@
+export { CheckboxLabel as default } from "./CheckboxLabel";
+export type { CheckboxLabelProps } from "./CheckboxLabel";