+++ src/index.ts
@@ -2,6 +2,8 @@ export { default as Banner } from "./components/Banner";
 export { default as Button } from "./components/Button";
 export { default as ButtonGroup } from "./components/ButtonGroup";
 export { default as Checkbox } from "./components/Checkbox";
+export { default as CheckboxInput } from "./components/CheckboxInput";
+export { default as CheckboxLabel } from "./components/CheckboxLabel";
 export { default as ClickableStyle } from "./components/ClickableStyle";
 export { default as Dropdown } from "./components/Dropdown";
 export { default as DropdownButton } from "./components/DropdownButton";