+++ CHANGELOG.md
@@ -2,14 +2,22 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+## [0.16.0](https://github.com/chanzuckerberg/edu-design-system/compare/v0.15.0...v0.16.0) (2022-05-03)
+
+### **Checkbox:** ([#1008](https://github.com/chanzuckerberg/edu-design-system/issues/1008))
+
+- `CheckboxInput` and `CheckboxLabel` have been moved to separate component folders
+- renamed `CheckboxLabel`'s `children` prop to `text`
+- renamed `size` options from "small" and "medium" to "sm" and "md" (respectively)
+
 ## [0.15.0](https://github.com/chanzuckerberg/edu-design-system/compare/v0.14.0...v0.15.0) (2022-05-02)
 
-### **Toast:** ([#1013](https://github.com/chanzuckerberg/edu-design-system/issues/1130))
+### **Toast:** ([#1013](https://github.com/chanzuckerberg/edu-design-system/issues/1013))
 
 - renamed `color` prop to `variant`
 - renamed "alert" color/variant to "error"
 
-### **Text:** ([#1025](https://github.com/chanzuckerberg/edu-design-system/issues/1250))
+### **Text:** ([#1025](https://github.com/chanzuckerberg/edu-design-system/issues/1025))
 
 - make `size` "lg" normal weight